[MOD-11717] (2/3) Implement array variants of Rust RsValue (#7280)
* Implement array variant of RsValue

Based on the `RsValueMap`, implement the array variant of `RsValueInternal`. Renames the `map` module in `value_ffi` and `value` to `collection`.

* Configure packed layout on RsValueCollection

In order to reduce the memory size of `RsValue` and
`SharedRsValue`, this commit annotates `RsValueCollection` with `#[repr(C, packed]` and updates the cbindgen configuration such that this gets reflected in the generated C header.

* Address review comments