[MOD-13100] improve handling of specDict_g and SchemaPrefixes_g global variables (#7807)
* init specDict_g to NULL

Will allow to check if it has been init or not when freeing global
memory in Rust tests.

This also prevent re-creating it multiple times leading to leaked
memory when running tests in parallel.

* init SchemaPrefixes_g to NULL

Will allow to check if it has been init or not when freeing global
memory in Rust tests.

This also prevent re-creating it multiple times leading to leaked
memory when running tests in parallel.